only error message logging for non-debug builds
made some changes to the tracing module so that in non-debug builds only
error messages will be logged.
modified make.py to show which mode (build or release) is being build
for.
while building in debug mode make commands verbosity option is set. so,
that it is easier to figure out which compiler flags are being passed.